Update reducer naming, add action types
