+++ app/constants/actionTypes.js
@@ -1,0 +1,2 @@
+export const RECEIVE_CURRENT_USER = 'RECEIVE_CURRENT_USER';
+export const CLEAR_CURRENT_USER = 'CLEAR_CURRENT_USER';
