+++ app/reducers/entities.js
@@ -1,7 +1,7 @@
 import { combineReducers } from 'redux';
 
-import users from './users_reducer';
-import posts from './posts_reducer';
+import users from './users';
+import posts from './posts';
 
 const entitiesReducer = combineReducers({
   users,
