--- app/reducers/root_reducer.js
@@ -1,15 +1,0 @@
-import { combineReducers } from 'redux';
-
-import session from './session_reducer';
-import errors from './errors_reducer';
-import ui from './ui_reducer';
-import entities from './entities_reducer';
-
-const rootReducer = combineReducers({
-  session,
-  errors,
-  ui,
-  entities
-});
-
-export default rootReducer;
